+++ doomtst/util/time.h
@@ -5,34 +5,34 @@
 #pragma once
 namespace timename {
 	struct time {
-		float value;
+		double value;
 
-		explicit time(float val) : value(val) {}
+		explicit time(double val) : value(val) {}
 		time() 
 			{
 				value = 0;
 			}
-		float dist(const time& oth) const {
+		double dist(const time& oth) const {
 			return std::fabs(oth.value - value);
 		}
 
 		bool operator<(const time& oth) const { return value < oth.value; }
 		bool operator==(const time& oth) const { return value == oth.value; }
 
-		time operator+(float offset) const { return time(value + offset); }
-		time operator-(float offset) const { return time(value - offset); }
+		time operator+(double offset) const { return time(value + offset); }
+		time operator-(double offset) const { return time(value - offset); }
 
-		float operator-(time offset) const { return (value - offset.value); }
+		double operator-(time offset) const { return (value - offset.value); }
 
 	};
 	struct duration;
 	struct TimeManager
 	{
-		float real_dt;
+		double real_dt;
 
-		float dt;
-		float smooth_fps;
-		float ElapsedTime;
+		double dt;
+		double smooth_fps;
+		double ElapsedTime;
 
 		TimeManager() {
 			
@@ -46,22 +46,22 @@ namespace timename {
 		
 		void calcfps() {
 
-			float CurrentTime= glfwGetTime();
+			double CurrentTime= glfwGetTime();
 			 real_dt = CurrentTime - ElapsedTime;
 			ElapsedTime = CurrentTime;
 			
 
-			float fps = 1.f / real_dt;
-			float inter_rate = 1;
+			double fps = 1.f / real_dt;
+			double inter_rate = 1;
 			dt = clamp(real_dt,0.f, 1.0f / min_frames);
 			
 			
-			float fps_change_rate = .2;
+			double fps_change_rate = .2;
 			fps_counter += real_dt;
 			if (fps_change_rate <fps_counter)
 			{
 				fps_counter = 0;
-				float fps_inter_rate =1;
+				double fps_inter_rate =1;
 				smooth_fps = lerp(smooth_fps, fps, fps_inter_rate);
 
 			}
@@ -73,7 +73,7 @@ namespace timename {
 		}
 		
 	private:
-		float fps_counter;
+		double fps_counter;
 		const int min_frames = 20;
 	};
 
@@ -87,13 +87,13 @@ namespace timename {
 		duration() :tm(nullptr), active(false) {
 
 		}
-		duration(float wait, TimeManager* tman):tm(tman),active(true) {
+		duration(double wait, TimeManager* tman):tm(tman),active(true) {
 			set(wait);
 		}
 		duration(TimeManager* tman) :tm(tman), active(false) {
 			
 		}
-		float remaining() {
+		double remaining() {
 			if (!active)
 			{
 				throw std::logic_error("cannot compute the remaining time of a duration");
@@ -107,7 +107,7 @@ namespace timename {
 		void disable() {
 			active = false;
 		}
-		void set(float dur) {
+		void set(double dur) {
 			if (dur<=0)
 			{
 				throw std::logic_error("duration must wasit for a positive number of time");
